k3s: drop 1.24 & 1.25 for 23.05
In-line with the policy described
[here](https://github.com/NixOS/nixpkgs/blob/30b82a186bc585872624a298a5169d1d237ce6a4/pkgs/applications/networking/cluster/k3s/README.md#versions-in-nixos-releases)
(xref #224483), drop versions of k3s that will not be supported for the
full duration of the NixOS release.

Since 22.11 has k3s 1.25, that means we must have k3s 1.26 at least.

Both k3s 1.24 and 1.25 will lose support before the 23.11 nixos release
goes out of support, so we should drop them. Respectively, 1.24 loses
support in July 2023, and 1.25 loses support in October 2023. NixOS is
supported through December 2023.